fix: move action folder
