+++ components/sidebarItem.tsx
@@ -11,7 +11,7 @@ import {
   PopoverTrigger,
   PopoverContent,
 } from "@/components/ui/popover";
-import { deleteExamById } from "@/actions/exam.action";
+import { deleteExamById } from "@/app/actions/exam.action";
 
 const SidebarItem = ({ label, examId }: { label: string; examId: string }) => {
   const [isOpen, setIsOpen] = useState(false);
